+++ samplesheets/assayapps/nucleotide_sequencing/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class NucleotideSequencingConfig(AppConfig):
+    name = 'samplesheets.assayapps.nucleotide_sequencing'